Refactor ToneSource into timeline and asset loader modules
